DOM data updated, loaded with OOTB code
The server passes down some JSON in a script tag inside the
div#taskboard.  This is the initial data that is used to populate the
task board.  This checkin updates that JSON before the page uses it so
that we can add the Blocked field and to each work item in that JSON.

This causes our custom data to be loaded into the standard data model
for the page.

Additionally, the custom renderer that we have written that helps to
draw the colors (_updateTileColors) now uses the Blocked field that
exists in the standard data model.